Redesigned some feeder implementation
