+++ src/gatling/java/uk/gov/hmcts/juror/performance/Util.java
@@ -25,8 +25,6 @@ public class Util {
     private static final AtomicInteger COUNTER;
     public static final Map<String, String> COMMON_HEADERS;
     public static final Map<String, String> PASSWORD_MAP;
-    public static final List<String> OWNER_LIST;
-
     public static final DateTimeFormatter STANDARD_DATE_FORMATTER = DateTimeFormatter.ofPattern("dd/MM/yyyy");
 
     private static final LocalDateGeneratorImpl LOCAL_DATE_GENERATOR = new LocalDateGeneratorImpl(
@@ -52,11 +50,6 @@ public class Util {
         PASSWORD_MAP = Map.of(
             "5BAA61E4C9B93F3F", "password"
         );
-        OWNER_LIST = Util.jdbcFeeder("select distinct owner from juror_mod.juror_pool")
-            .readRecords()
-            .stream()
-            .map(stringObjectMap -> String.valueOf(stringObjectMap.get("owner")))
-            .toList();
     }
 
     public static CheckBuilder.Final saveCsrf() {
@@ -70,24 +63,6 @@ public class Util {
 
     public static CheckBuilder.Final validateHeading(String headingValue) {
         return css("h1.govuk-heading-l").is(headingValue);
-    }
-
-    public static FeederBuilder<Object> jdbcFeeder(String sql) {
-        log.info(
-            "Creating jdbcFeeder: " + Config.DB_URL + ", " + Config.DB_USERNAME + ", " + Config.DB_PASSWORD + ", "
-                + sql);
-
-        return JdbcDsl.jdbcFeeder(Config.DB_URL, Config.DB_USERNAME, Config.DB_PASSWORD, sql);
-    }
-
-    public static FeederBuilder<Object> listFeeder(String key, List<Object> items) {
-        return CoreDsl.listFeeder(items.stream()
-            .map(item -> Map.of(key, item)).toList()
-        );
-    }
-
-    public static FeederBuilder<Object> listFeeder(List<Map<String, Object>> items) {
-        return CoreDsl.listFeeder(items);
     }
 
     public static String getPasswordFromDB(String password) {
